Move EventAndTriggers OptionTag into its own file
This breaks a non-enforced dependency of libGaugeSourceFunctions upon
module_ConstGlobalCache, which can cause the build to fail.
